Refactor import paths and file names general helpers
- the previously used namings for these files with double suffixes (helper.colors.ts) were causing trouble with Tailwind unfortunately, so I renamed them
