+++ lib/helpers/arrays.ts
